twitter-abm-viral/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
plotly>=5.3.0 